# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Ticket Zetu application",
        "title": "Ticket Zetu API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@ticketzetu.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/api/v1",
    "basePath": "/",
    "paths": {
...
        "/users/me/preferences": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user preferences",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserPreferencesDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the authenticated user's preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user preferences",
                "parameters": [
                    {
                        "description": "Preferences to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserPreferencesDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserPreferencesDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.UpdateUserPreferencesDto": {
            "type": "object",
            "properties": {
                "allow_following": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "show_email": {
                    "type": "boolean"
                },
                "show_gender": {
                    "type": "boolean"
                },
                "show_location": {
                    "type": "boolean"
                },
                "show_phone": {
                    "type": "boolean"
                },
                "show_profile": {
                    "type": "boolean"
                },
                "show_role": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string",
                    "enum": [
                        "light",
                        "dark"
                    ]
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserPreferencesDto": {
            "type": "object",
            "properties": {
                "allow_following": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "show_email": {
                    "type": "boolean"
                },
                "show_gender": {
                    "type": "boolean"
                },
                "show_location": {
                    "type": "boolean"
                },
                "show_phone": {
                    "type": "boolean"
                },
                "show_profile": {
                    "type": "boolean"
                },
                "show_role": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}